apps/fm-app/pyproject.toml
[project]
requires-python = ">= 3.13"
name = "flow-manager"
version = "0.1.0"
dependencies = [
    "alembic==1.14.0",
    "amqp==5.3.0",
    "annotated-types==0.7.0",
    "anthropic==0.47.2",
    "anyio==4.6.2.post1",
    "asynch==0.2.5",
    "asyncpg==0.30.0",
    "billiard==4.2.1",
    "cachetools==5.5.0",
    "celery==5.4.0",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.4.0",
    "ciso8601==2.3.2",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "clickhouse-connect==0.8.16",
    "clickhouse-driver==0.2.9",
    "clickhouse-sqlalchemy==0.3.2",
    "cryptography==43.0.3",
    "dash==2.13.0",
    "distro==1.9.0",
    "dnspython==2.7.0",
    "docstring-parser==0.16",
    "duckdb>=1.2.2",
    "email-validator==2.2.0",
    "fastapi==0.115.4",
    "fastapi-cli==0.0.5",
    "fastmcp==2.2.7",
    "google-api-core==2.23.0",
    "google-auth==2.36.0",
    "google-cloud-aiplatform==1.71.1",
    "google-cloud-bigquery==3.27.0",
    "google-cloud-core==2.4.1",
    "google-cloud-resource-manager==1.13.1",
    "google-cloud-storage==2.18.2",
    "google-crc32c==1.6.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.66.0",
    "greenlet==3.1.1",
    "grpc-google-iam-v1==0.13.1",
    "grpcio==1.67.1",
    "grpcio-status==1.67.1",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.4",
    "jiter==0.8.0",
    "jsonschema>=4.25.0",
    "kaleido==0.4.2",
    "kombu==5.4.2",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "leb128==1.0.8",
    "lz4==4.3.3",
    "mako==1.3.6",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.9.4",
    "mdurl==0.1.2",
    "numpy==2.1.3",
    "openai==1.76.2",
    "openai-agents>=0.0.14",
    "packaging==24.2",
    "pandas>=2.2.3",
    "plotly==5.24.1",
    "prompt-toolkit==3.0.48",
    "proto-plus==1.25.0",
    "protobuf==5.28.3",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pygments==2.18.0",
    "pyjwt==2.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.17",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "rich==13.9.4",
    "rsa==4.9",
    "shapely==2.0.6",
    "shellingham==1.5.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.36",
    "sqlglot==26.14.0",
    "sqlglotrs==0.4.0",
    "starlette==0.41.2",
    "structlog==24.4.0",
    "tqdm==4.67.1",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "tzlocal==5.2",
    "urllib3==2.2.3",
    "uuid7==0.1.0",
    "uvicorn==0.32.0",
    "uvloop==0.21.0",
    "vertexai==1.71.1",
    "vine==5.1.0",
    "watchdog==6.0.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "zstd==1.5.5.1",
]

[build-system]
requires = ["setuptools"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["alembic", "bulk_test", "k8s"]

[tool.ruff]
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "alembic",
    "bulk_test",
    "prompt_elements.py"
]
line-length = 88

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
select = ["E", "F", "I"]

[tool.black]
line-length = 88
target-version = ['py312']  # Adjust based on your Python versions

[tool.isort]
profile = "black"
line_length = 88

[tool.uv.workspace]
members = [
    "bulk_test",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "ruff>=0.11.6",
]
